+++ packages/plugin-coverage/src/lib/coverage-plugin.ts
@@ -5,12 +5,14 @@ import {
   type PluginConfig,
   validate,
 } from '@code-pushup/models';
-import { capitalize } from '@code-pushup/utils';
+import { logger, pluralizeToken } from '@code-pushup/utils';
 import {
   type CoveragePluginConfig,
   type CoverageType,
   coveragePluginConfigSchema,
 } from './config.js';
+import { COVERAGE_PLUGIN_SLUG, COVERAGE_PLUGIN_TITLE } from './constants.js';
+import { formatMetaLog, typeToAuditSlug, typeToAuditTitle } from './format.js';
 import { createRunnerFunction } from './runner/runner.js';
 import { coverageDescription, coverageTypeWeightMapper } from './utils.js';
 
@@ -39,8 +41,8 @@ export async function coveragePlugin(
 
   const audits = coverageConfig.coverageTypes.map(
     (type): Audit => ({
-      slug: `${type}-coverage`,
-      title: `${capitalize(type)} coverage`,
+      slug: typeToAuditSlug(type),
+      title: typeToAuditTitle(type),
       description: coverageDescription[type],
     }),
   );
@@ -58,6 +60,12 @@ export async function coveragePlugin(
     })),
   };
 
+  logger.info(
+    formatMetaLog(
+      `Created ${pluralizeToken('audit', audits.length)} (${coverageConfig.coverageTypes.join('/')} coverage) and 1 group`,
+    ),
+  );
+
   const packageJson = createRequire(import.meta.url)(
     '../../package.json',
   ) as typeof import('../../package.json');
@@ -65,8 +73,8 @@ export async function coveragePlugin(
   const scoreTargets = coverageConfig.scoreTargets;
 
   return {
-    slug: 'coverage',
-    title: 'Code coverage',
+    slug: COVERAGE_PLUGIN_SLUG,
+    title: COVERAGE_PLUGIN_TITLE,
     icon: 'folder-coverage-open',
     description: 'Official Code PushUp code coverage plugin.',
     docsUrl: 'https://www.npmjs.com/package/@code-pushup/coverage-plugin/',