+++ packages/plugin-lighthouse/src/lib/format.ts
@@ -1,0 +1,4 @@
+import { pluginMetaLogFormatter } from '@code-pushup/utils';
+import { LIGHTHOUSE_PLUGIN_TITLE } from './constants.js';
+
+export const formatMetaLog = pluginMetaLogFormatter(LIGHTHOUSE_PLUGIN_TITLE);