+++ packages/plugin-typescript/src/lib/runner/ts-runner.int.test.ts
@@ -2,12 +2,12 @@ import { describe, expect } from 'vitest';
 import { getTypeScriptDiagnostics } from './ts-runner.js';
 
 describe('getTypeScriptDiagnostics', () => {
-  it('should return valid diagnostics', async () => {
-    await expect(
+  it('should return valid diagnostics', () => {
+    expect(
       getTypeScriptDiagnostics({
         tsconfig:
           'packages/plugin-typescript/mocks/fixtures/basic-setup/tsconfig.json',
       }),
-    ).resolves.toHaveLength(8);
+    ).toHaveLength(5);
   });
 });