+++ packages/utils/src/lib/file-system.ts
@@ -180,3 +180,46 @@ export function splitFilePath(filePath: string): SplitFilePath {
   }
   return { folders, file };
 }
+
+export function truncatePaths(paths: string[]): string[] {
+  const segmentedPaths = paths
+    .map(splitFilePath)
+    .map(({ folders, file }): string[] => [...folders, file]);
+
+  const first = segmentedPaths[0];
+  const others = segmentedPaths.slice(1);
+  if (!first) {
+    return paths;
+  }
+
+  /* eslint-disable functional/no-let,functional/no-loop-statements,unicorn/no-for-loop */
+  let offsetLeft = 0;
+  let offsetRight = 0;
+  for (let left = 0; left < first.length; left++) {
+    if (others.every(segments => segments[left] === first[left])) {
+      offsetLeft++;
+    } else {
+      break;
+    }
+  }
+  for (let right = 1; right <= first.length; right++) {
+    if (others.every(segments => segments.at(-right) === first.at(-right))) {
+      offsetRight++;
+    } else {
+      break;
+    }
+  }
+  /* eslint-enable functional/no-let,functional/no-loop-statements,unicorn/no-for-loop */
+
+  return segmentedPaths.map(segments => {
+    const uniqueSegments = segments.slice(
+      offsetLeft,
+      offsetRight > 0 ? -offsetRight : undefined,
+    );
+    return path.join(
+      offsetLeft > 0 ? '…' : '',
+      ...uniqueSegments,
+      offsetRight > 0 ? '…' : '',
+    );
+  });
+}